+++ src/main/java/org/JavaArt/TicketManager/service/SectorService.java
@@ -17,12 +17,12 @@ public class SectorService {
         return sectorRepository.getSectorsByEvent(event);
     }
 
-    public Map<Double,List<Sector>> getSectorsByEventSortedByPrice(Event event) {
+    public Map<Double, List<Sector>> getSectorsByEventSortedByPrice(Event event) {
         List<Sector> sectors = getSectorsByEvent(event);
-        Map<Double,List<Sector>> sectorsMap = new TreeMap<>();
-        for (Sector sector : sectors){
+        Map<Double, List<Sector>> sectorsMap = new TreeMap<>();
+        for (Sector sector : sectors) {
             List<Sector> orderedSectors = sectorsMap.get(sector.getPrice());
-            if (orderedSectors==null) {
+            if (orderedSectors == null) {
                 orderedSectors = new ArrayList<>();
                 orderedSectors.add(sector);
                 sectorsMap.put(sector.getPrice(), orderedSectors);
@@ -37,7 +37,7 @@ public class SectorService {
         return sectorRepository.getSectorsByEventOrderPrice(event);
     }
 
-    public Sector getSectorById(int id) {
+    public Sector getSectorById(Long id) {
         return sectorRepository.getSectorById(id);
     }
 
@@ -46,22 +46,22 @@ public class SectorService {
         List<String> legenda = new LinkedList<>();
         StringBuilder buf = new StringBuilder(100);
         int index;
-        int element=0;
+        int element = 0;
         sortByPrice.add(0, sectorByPrice.get(0).getPrice());
         for (int i = 1; i < sectorByPrice.size(); i++) {
             if (sectorByPrice.get(i).getPrice() > sectorByPrice.get(i - 1).getPrice())
                 sortByPrice.add(sectorByPrice.get(i).getPrice());
         }
-        List<Sector> sector=sectorRepository.getSectorsByEvent(sectorByPrice.get(0).getEvent());
+        List<Sector> sector = sectorRepository.getSectorsByEvent(sectorByPrice.get(0).getEvent());
         for (int j = 0; j < sortByPrice.size(); j++) {
             buf.append(sortByPrice.get(j)).append(" грн.  Сектор ");
             index = 0;
-            for (Sector sector2: sector) {
+            for (Sector sector2 : sector) {
                 if ((double) sector2.getPrice() == sortByPrice.get(j)) {
                     if (index == 0) buf.append(sector2.getName());
                     if (index > 0) buf.append(", ").append(sector2.getName());
                     index++;
-                    if (buf.toString().length()>46){
+                    if (buf.toString().length() > 46) {
                         legenda.add(element, buf.toString());
                         buf.delete(0, 99);
                         buf.append(sortByPrice.get(j)).append(" грн.  Сектор ");
@@ -70,8 +70,8 @@ public class SectorService {
                     }
                 }
             }
-            if (buf.toString().length()<15)    buf.delete(0, 99);
-            else{
+            if (buf.toString().length() < 15) buf.delete(0, 99);
+            else {
                 legenda.add(element, buf.toString());
                 buf.delete(0, 99);
                 element++;
